+++ include/Event.h
@@ -1,0 +1,12 @@
+#ifndef EVENT_H
+#define EVENT_H
+
+class Event {
+private :
+
+public:
+    Event();
+    ~Event();
+};
+
+#endif