+++ include/Item.h
@@ -2,6 +2,7 @@
 #define ITEM_H
 
 #include <string>
+#include <iostream>
 #include "RARITY.h"
 
 class Item {
@@ -21,7 +22,13 @@ public:
     //Functions
     std::string toString();
 
-};
+    // Accessors
+    inline unsigned long getId() const { return this->id; }
+    inline std::string getName() const { return this->name; }
+    inline std::string getDesc() const { return this->desc; }
+    inline int getPrice() const { return this->price; }
+    inline RARITY getRarity() const { return this->rarity; }
 
+};
 
 #endif