+++ src/Event.cpp
@@ -1,0 +1,10 @@
+#include "../include/Event.h"
+
+Event::Event()
+{
+
+}
+
+Event::~Event()
+{
+}